Move hibernate packages around
