+++ src/main/java/learn_to_code/frameworks/hibernate/foreign_key_example/HibernateRunner.java
@@ -1,0 +1,7 @@
+package learn_to_code.frameworks.hibernate.foreign_key_example;
+
+public class HibernateRunner {
+    public static void main(String[] args) {
+        //TODO: do something!
+    }
+}
