+++ src/main/java/learn_to_code/frameworks/hibernate/one_table_example/HibernateRunner.java
@@ -1,7 +1,13 @@
-package learn_to_code.frameworks.hibernate;
+package learn_to_code.frameworks.hibernate.one_table_example;
 
+import learn_to_code.frameworks.hibernate.HibernateUtil;
 import org.hibernate.Session;
 
+/**
+ * Example of populating a table with data.
+ *
+ * Note that session object implements auto-closable, so using try-with-resources is encouraged.
+ */
 public class HibernateRunner {
     public static void main(String[] args) {
         try (Session session = HibernateUtil.getSessionFactory().openSession()) {
@@ -14,8 +20,6 @@ public class HibernateRunner {
 
             session.createQuery("from Town").list().forEach(c -> System.out.println(c));
             session.getTransaction().commit();
-
-            session.close();
         }
     }
 }
