+++ src/main/resources/hibernate.cfg.xml
@@ -14,7 +14,7 @@
         <property name="show_sql">true</property>
         <property name="hbm2ddl.auto">create</property>
 
-        <mapping class="learn_to_code.frameworks.hibernate.Town"/>
+        <mapping class="learn_to_code.frameworks.hibernate.one_table_example.Town"/>
 
     </session-factory>
 </hibernate-configuration>
